Convert to Python3 compatible code
Change most of the code to make the package work with a Python 3 interpreter.  This mostly involves changing print statements to be inside parentheses, but also includes changing how the package imports modules from within itself in the form of adding periods before the module names.  Also cPickle is not an option in Python 3 so import vanilla pickle in operation.py