Fix package installation with setuptools
Add setup() metadata that tells the installer to keep the import .txt and .prm files
for atomtype and id parsing.  This was the problem with the setup.py install and it
now appears to be fixed, change the README back accordingly.
Remove param/amber/__init__.py as it was obsolete.  Give absolute path in amber/parser.py
because it was being called in setup.py.  Change package directory function call in the base __init__.py.  Add .npy files in .gitignore,
need to be wary of this later.