Correct the displacement calculation
Make the necessary changes for 'correct' displacement calculations, which were addressed in #30.  The problem was in the integration.  Keep the new code in 'greens_func.py' while the old method is kept in 'greens_func_old.py'.  Add functionality to plot the displacements of every atom simultaneously.  Delete 'elementwise_gf.py' as that existed to test the construction of arrays by doing so element-wise which ended up not being the problem in the end.