Add Interaction descriptor, remove default ff
At minimum, add the Interaction class to be used as a descriptor for the interaction attributes of Forcefield.  The 'state variable' code hasn't been added yet.
Also remove the use of default forcefields.  In the basic workflow of kappa, it is probably better that the user explicitly specify a forcefield ALL THE TIME.