Finish 3D Green's function calculation for 1d chain
Finish the 3D GF calculation and plotting, assuming 3D harmonic oscillators (kr2).  Also remove the factor of 2 in the GF coefficient calculation.  We believe a factor goes away from the derivative of the Dirac delta function in the notes.  I believe all the removal does is change the scale of the solutions.