Add attempted thermal conductivity calculation
Add the analytical thermal conductivity calculation for the simple 1d chain in 3d.  However the calculated value comes out to precisely 0.  Need to refer to the notes to catch mistakes.  Put in code in the 3D GF file to check to make sure the coefficients and eigenvectors match up.  Clean up code in molecule.py and operation.py.