Use interfaces in chaining function
Incorporate the interfaces of molecules when they are chained together.  This involves lining up the molecules so that the interface norms are anti-parallel. Add an invert method to molecule objects to account for the cases where norms are parallel.  Add normalization code to interface norms.  At this point it is assumed that atoms can only belong to one interface if any.