Add interface tracking
Add an interface tracking attribute to Molecule, used to track which interfacethe listed atoms are attached to.  Combine facetrack in operation._combine.  This will go towards #42.  Need to combine this with interface paths so that when the user is measuring thermal conductivity he/she pairs 2 interfaces and the calculate considers the 'wiggle' of all the atoms beyond the specified interfaces, searching through the interface paths with the help of Molecule.facetrack.