Change indices of imine atoms
In an effort towards some standardization, change the indices of the imine atoms such that its interfaces are at 0 and 1 instead of 3,4.  For simple 2-interface molecules like this it would be good to have a standard 0,1 atomic interfaces.  Make the required changes in the imine lattice generation, imine builder, and imine chain builder.