Remove drawing of face.norm if it doesn't go in the x-y plane
In plot.bonds, prevent a 'soft' error that would occur if the attempted quiver object is attempted to be drawn of length zero.  Add an if check the length in the x,y plane is less than some small number.  try/except block wouldn't catch this...it's not a hard interrupt error.