Add checks for validity of chain,_combine parameters
Add the explicit checks for the validity of the number of molecules in molList and number of index pairs for operation.chain, combining atoms of the same atomic number in operation._combine, combining atoms that are not already closed in _combine.  There might be a better way than using if blocks (structure it for try/except blocks?) but it's here now.  Fixes #47.  The chain code is still ugly though.