Add finder of interface crossing interactions
Add the code to find the interactions that cross the interfaces then add
their driving power together.  Still has some kinks to work out but should
be suitable for a test calculation.  Right now many things are assumed in
the code such as the interfaces surrounding the 'root' molecule
(facetrack=1)