Fix backtracking line search; eliminate bad steps
Correct the backtracking line search algorithm such that the atoms' positions reset when the the test energy is calculated.  Also fix a critical error in the conjugate gradient function in which the step taken wasn't actually the step used to calculate the step size.  Also remove the changing of the 'e' state variable.  Properly unpack the gradient.  Scale the initial step guess with the previous stepsize.