Incorporate atom type file from AmberTools
Remove old atom type files copy & pasted from antechamber paper.  Remove the tabs from them to get the csv.reader to parse the lines.  Remove print statements in Molecule._configure.  Give a warning instead of a print statment in a DU atom is assigned.  Actually discovered that the previous force constants were all of the 'CZ' type!  Found that and squashed it.