Add ModeInspector class
which allows users to look at certain properties of molecules.  For the time being it will plot the participation ratios.  This can replace the now obsolete inspection functions.  Also import everything from kappa.conductivity to the package level because the functions that do the actual calculating moved to a new package.